Permission added and menu issue fixed and dashboard redirect to users

See merge request sudippalash/go2fan-script!18

+++ app/Http/Controllers/Panel/Appearance/FileController.php
@@ -13,52 +13,63 @@ class FileController extends Controller
 {
     public function index()
     {
-        $data = File::where('panel_id', Auth::user()->panel_id)->orderBy('id', 'DESC')->get();
-        return view('panel.appearance.file', compact('data'));
+        if (Auth::user()->can('files')) {
+            $data = File::where('panel_id', Auth::user()->panel_id)->orderBy('id', 'DESC')->get();
+            return view('panel.appearance.file', compact('data'));
+        }else{
+            return view('panel.permission');
+        }
     }
 
     public function store(Request $request)
     {
-        if ($request->has('files')){
-            $data = [];
-            foreach ($request->file('files') as $file){
-                $size = $file->getSize();
-                $mimeType = $file->getMimeType();
-                $ext  = $file->getClientOriginalExtension();
-                $mainName  = $file->getClientOriginalName();
-                $fileName = uniqid('file-').'.'.$ext;
-                $url = asset('storage/files/'.$fileName);
-                $file = Image::make($file);
-                $data[] = [
-                    'panel_id'   => Auth::user()->panel_id,
-                    'name'       => $mainName,
-                    'size'       => $size,
-                    'mime_type'  => $mimeType,
-                    'extension'  => $ext,
-                    'url'        => $url,
-                    'created_by' => Auth::user()->id,
-                ];
-                Storage::disk('public')->put("files/".$fileName, (string) $file->encode());
+        if (Auth::user()->can('files')) {
+            if ($request->has('files')){
+                $data = [];
+                foreach ($request->file('files') as $file){
+                    $size = $file->getSize();
+                    $mimeType = $file->getMimeType();
+                    $ext  = $file->getClientOriginalExtension();
+                    $mainName  = $file->getClientOriginalName();
+                    $fileName = uniqid('file-').'.'.$ext;
+                    $url = asset('storage/files/'.$fileName);
+                    $file = Image::make($file);
+                    $data[] = [
+                        'panel_id'   => Auth::user()->panel_id,
+                        'name'       => $mainName,
+                        'size'       => $size,
+                        'mime_type'  => $mimeType,
+                        'extension'  => $ext,
+                        'url'        => $url,
+                        'created_by' => Auth::user()->id,
+                    ];
+                    Storage::disk('public')->put("files/".$fileName, (string) $file->encode());
+                }
             }
-        }
 
-        if(!empty($data)){
-            File::insert($data);
-            return redirect()->back()->with('success', 'File save successfully !!');
+            if(!empty($data)){
+                File::insert($data);
+                return redirect()->back()->with('success', 'File save successfully !!');
+            }else{
+                return redirect()->back()->with('error', "File con't save ! Please try again !!");
+            }
         }else{
-            return redirect()->back()->with('error', "File con't save ! Please try again !!");
+            return view('panel.permission');
         }
 
     }
 
     public function destroy($id)
     {
-
-        $data = File::where('panel_id', Auth::user()->panel_id)->where('id', $id)->first();
-        if (empty($data)){
-            return redirect()->back()->with('error', "This file can't deleted! Please try again !!");
+        if (Auth::user()->can('files')) {
+            $data = File::where('panel_id', Auth::user()->panel_id)->where('id', $id)->first();
+            if (empty($data)){
+                return redirect()->back()->with('error', "This file can't deleted! Please try again !!");
+            }
+            $data->delete();
+            return redirect()->back()->with('success', 'This file delete successfully !!');
+        }else{
+            return view('panel.permission');
         }
-        $data->delete();
-        return redirect()->back()->with('success', 'This file delete successfully !!');
     }
 }
